requirements.txt
streamlit
google-generativeai
PyMuPDF
pytesseract
pillow
beautifulsoup4
requests
python-dotenv
